Update to 1.21.6-rc1

+++ src/main/java/de/maxhenkel/easypiglins/gui/InputOutputScreen.java
@@ -26,12 +26,12 @@ public abstract class InputOutputScreen<T extends AbstractContainerMenu> extends
     protected void renderLabels(GuiGraphics guiGraphics, int x, int y) {
         drawCentered(guiGraphics, getTopText(), 9, FONT_COLOR);
         drawCentered(guiGraphics, getBottomText(), 40, FONT_COLOR);
-        guiGraphics.drawString(font, playerInventory.getDisplayName().getVisualOrderText(), 8F, (float) (imageHeight - 96 + 3), FONT_COLOR, false);
+        guiGraphics.drawString(font, playerInventory.getDisplayName().getVisualOrderText(), 8, imageHeight - 96 + 3, FONT_COLOR, false);
     }
 
     protected void drawCentered(GuiGraphics guiGraphics, MutableComponent text, int y, int color) {
         int width = font.width(text);
-        guiGraphics.drawString(font, text.getVisualOrderText(), imageWidth / 2F - width / 2F, y, color, false);
+        guiGraphics.drawString(font, text.getVisualOrderText(), imageWidth / 2 - width / 2, y, color, false);
     }
 
     protected abstract MutableComponent getTopText();
